requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
japanize-matplotlib>=1.1.3
prophet>=1.1.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
torch>=2.0.0
chronos-forecasting>=1.0.0
